Simplify Entry structure
